+++ .gitignore
@@ -6,7 +6,7 @@ bin/
 *.tfplan
 
 # Config file might have secrets
-.rover.*
+.rover.yaml
 
 # Goreleaser
 dist/
