+++ cmd/ci.go
@@ -8,36 +8,12 @@ package cmd
 
 import (
 	"fmt"
-	"io/ioutil"
-	"os"
 
-	"github.com/fatih/color"
+	"github.com/aztfmod/rover/pkg/console"
+	"github.com/aztfmod/rover/pkg/symphony"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
-	"gopkg.in/yaml.v2"
 )
-
-type SymphonyConfig struct {
-	Environment  string `yaml:"environment,omitempty"`
-	Repositories []struct {
-		Name   string `yaml:"name,omitempty"`
-		URI    string `yaml:"uri,omitempty"`
-		Branch string `yaml:"branch,omitempty"`
-	}
-	Levels []struct {
-		Level     string `yaml:"level,omitempty"`
-		Type      string `yaml:"type,omitempty"`
-		Launchpad bool   `yaml:"launchpad,omitempty"`
-		Stacks    []struct {
-			Stack             string `yaml:"stack,omitempty"`
-			LandingZonePath   string `yaml:"landingZonePath,omitempty"`
-			ConfigurationPath string `yaml:"configurationPath,omitempty"`
-			TfState           string `yaml:"tfState,omitempty"`
-		}
-	}
-}
-
-var symphonyConfig SymphonyConfig
 
 // ciCmd represents the ci command
 var ciCmd = &cobra.Command{
@@ -47,13 +23,13 @@ var ciCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 
 		symphonyConfigFileName, _ := cmd.Flags().GetString("symphony-config")
-		verbose, _ := cmd.Flags().GetBool("verbose")
+		debug, _ := cmd.Flags().GetBool("debug")
 
-		err := readAndUnmarshallConfig(symphonyConfigFileName)
+		symphonyConfig, err := symphony.NewSymphonyConfig(symphonyConfigFileName)
 		cobra.CheckErr(err)
 
-		if verbose {
-			outputVerbose(symphonyConfigFileName)
+		if debug {
+			symphonyConfig.OutputDebug(symphonyConfigFileName)
 		}
 
 		run(symphonyConfigFileName)
@@ -63,24 +39,7 @@ var ciCmd = &cobra.Command{
 func run(symphonyConfigFileName string) {
 	fmt.Println()
 
-	color.Red("Running CI command, config: %s", symphonyConfigFileName)
-}
-
-func readAndUnmarshallConfig(symphonyConfigFileName string) error {
-	reader, _ := os.Open(symphonyConfigFileName)
-	buf, _ := ioutil.ReadAll(reader)
-	err := yaml.Unmarshal(buf, &symphonyConfig)
-
-	return err
-}
-
-func outputVerbose(symphonyConfigFileName string) {
-	fmt.Println()
-
-	color.Blue("Verbose output of %s", symphonyConfigFileName)
-	color.Green(" - Environment: %s", symphonyConfig.Environment)
-	color.Green(" - Number of repositories: %d", len(symphonyConfig.Repositories))
-	color.Green(" - Number of levels: %d", len(symphonyConfig.Levels))
+	console.Infof("Running CI command, config: %s\n", symphonyConfigFileName)
 }
 
 func init() {
