+++ samples/sample.symphony.yaml
